requirements.txt
python == 3.10
pandas
numpy
os
glob
imutils
pillow
image